Improve all commands

+++ commands/leave.js
@@ -1,19 +1,18 @@
 // TODO: Implement broadcast functions like stop, pause, and resume
 
-const stop = (message, audioPlayer, broadcast) => {
+const leave = async (message, audioPlayer, broadcast) => {
+  if (!audioPlayer.voiceChannel) {
+    await message.delete();
+    await message.channel.send("```I'm not currently in any Voice Channel!```");
+  }
+
   try {
-    broadcast.end();
-    audioPlayer.voiceChannel.leave();
-    audioPlayer.status = 'Inactive';
-    message.delete();
-    message.channel.send(
-      `${message.author}\`\`\`Stopped Playing: ${
-        audioPlayer.currentAudio
-      }\`\`\``
-    );
+    require('./stop.js')(message, audioPlayer, broadcast);
+    await audioPlayer.voiceChannel.leave();
+    audioPlayer.voiceChannel = null;
   } catch (err) {
     console.log('Failed to leave.');
   }
 };
 
-module.exports = stop;
+module.exports = leave;
